Refactor MovieBox  component
