+++ package.json
@@ -7,6 +7,7 @@
     "@types/node": "12.11.7",
     "@types/react": "16.9.11",
     "@types/react-dom": "16.9.3",
+    "global": "^4.4.0",
     "node-sass": "^4.13.0",
     "react": "^16.11.0",
     "react-app-polyfill": "^1.0.4",
@@ -16,6 +17,7 @@
     "react-svg": "^10.0.23",
     "redux": "^4.0.4",
     "redux-thunk": "^2.3.0",
+    "surge": "^0.21.3",
     "typescript": "3.6.4"
   },
   "scripts": {
