--- src/components/organisms/MessageBox/MessageBox.scss
@@ -1,18 +1,0 @@
-@import "../../../config/variables.scss";
-
-.error-box {
-  margin: 0 auto;
-  max-width: $page-width-max;
-  padding: $main-padding * 3 $main-padding;
-  background: $color-bg-dark;
-  display: flex;
-  flex-direction: column;
-  justify-content: center;
-  align-items: center;
-  height: 100vh;
-  @media screen and (min-width: map-get($breakpoints, sm)) {
-    margin-top: $main-padding * 2;
-    padding: 0;
-    height: 50vh;
-  }
-}
